+++ react-complete-guide/src/utils/delay.js
@@ -1,0 +1,3 @@
+export default (timeout, cb) => {
+  return setTimeout(() => cb(), timeout);
+};